pyproject.toml
[project]
name = "associative-memory"
version = "0.1.0"
description = "Multi-layer associative memory system with query-conditioned retrieval"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.24",
    "networkx>=3.0",
    "httpx>=0.25",
    "msgpack>=1.0",
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "click>=8.1",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[project.scripts]
amem = "cli.main:cli"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends._legacy:_Backend"
